+++ pytest_reportportal/service.py
@@ -3,15 +3,18 @@ import logging
 import os.path
 import sys
 import threading
-from os import getenv, curdir
+from functools import wraps
+from os import curdir
 from time import time, sleep
-from typing import List
+from typing import List, Any, Optional, Set, Dict, Tuple
 
 from _pytest.doctest import DoctestItem
 from aenum import auto, Enum, unique
 from pytest import Class, Function, Module, Package, Item, Session, \
     PytestWarning
 from reportportal_client.core.rp_issues import Issue, ExternalIssue
+
+from .config import AgentConfig
 
 try:
     from pytest import Instance
@@ -28,14 +31,15 @@ from reportportal_client.helpers import (
 
 log = logging.getLogger(__name__)
 
-MAX_ITEM_NAME_LENGTH = 256
-TRUNCATION_STR = '...'
-ROOT_DIR = str(os.path.abspath(curdir))
-PYTEST_MARKS_IGNORE = {'parametrize', 'usefixtures', 'filterwarnings'}
-NOT_ISSUE = Issue('NOT_ISSUE')
-ISSUE_DESCRIPTION_LINE_TEMPLATE = '* {}:{}'
-ISSUE_DESCRIPTION_URL_TEMPLATE = ' [{issue_id}]({url})'
-ISSUE_DESCRIPTION_ID_TEMPLATE = ' {issue_id}'
+MAX_ITEM_NAME_LENGTH: int = 256
+TRUNCATION_STR: str = '...'
+ROOT_DIR: str = str(os.path.abspath(curdir))
+PYTEST_MARKS_IGNORE: Set[str] = {'parametrize', 'usefixtures',
+                                 'filterwarnings'}
+NOT_ISSUE: Issue = Issue('NOT_ISSUE')
+ISSUE_DESCRIPTION_LINE_TEMPLATE: str = '* {}:{}'
+ISSUE_DESCRIPTION_URL_TEMPLATE: str = ' [{issue_id}]({url})'
+ISSUE_DESCRIPTION_ID_TEMPLATE: str = ' {issue_id}'
 
 
 def timestamp():
@@ -43,12 +47,12 @@ def timestamp():
     return str(int(time() * 1000))
 
 
-def trim_docstring(docstring):
+def trim_docstring(docstring: str) -> str:
     """
     Convert docstring.
 
     :param docstring: input docstring
-    :return: docstring
+    :return: trimmed docstring
     """
     if not docstring:
         return ''
@@ -93,29 +97,52 @@ class ExecStatus(Enum):
     FINISHED = auto()
 
 
-class PyTestServiceClass(object):
+def check_rp_enabled(func):
+    """Verify is RP is enabled in config."""
+
+    @wraps(func)
+    def wrap(*args, **kwargs):
+        if args and isinstance(args[0], PyTestServiceClass):
+            if not args[0].rp:
+                return
+        func(*args, **kwargs)
+
+    return wrap
+
+
+class PyTestServiceClass:
     """Pytest service class for reporting test results to the Report Portal."""
 
-    def __init__(self, agent_config):
+    _config: AgentConfig
+    _issue_types: Dict[str, str]
+    _tree_path: Dict[Item, List[Dict[str, Any]]]
+    _log_levels: Tuple[str, str, str, str, str]
+    _start_tracker: Set[str]
+    _launch_id: Optional[str]
+    agent_name: str
+    agent_version: str
+    ignored_attributes: List[str]
+    parent_item_id: Optional[str]
+    rp: Optional[RPClient]
+    project_settings: Dict[str, Any]
+
+    def __init__(self, agent_config: AgentConfig) -> None:
         """Initialize instance attributes."""
         self._config = agent_config
         self._issue_types = {}
         self._tree_path = {}
         self._log_levels = ('TRACE', 'DEBUG', 'INFO', 'WARN', 'ERROR')
-        self._skip_analytics = getenv('AGENT_NO_ANALYTICS')
         self._start_tracker = set()
-        self._process_level_lock = threading.Lock()
         self._launch_id = None
         self.agent_name = 'pytest-reportportal'
         self.agent_version = get_package_version(self.agent_name)
         self.ignored_attributes = []
-        self.log_batch_size = 20
         self.parent_item_id = None
         self.rp = None
         self.project_settings = {}
 
     @property
-    def issue_types(self):
+    def issue_types(self) -> Dict[str, str]:
         """Issue types for the Report Portal project."""
         if not self._issue_types:
             if not self.project_settings:
@@ -152,14 +179,13 @@ class PyTestServiceClass(object):
         }
         return start_rq
 
-    def start_launch(self):
+    @check_rp_enabled
+    def start_launch(self) -> Optional[str]:
         """
         Launch test items.
 
         :return: item ID
         """
-        if self.rp is None:
-            return
         sl_pt = self._build_start_launch_rq()
         log.debug('ReportPortal - Start launch: request_body=%s', sl_pt)
         self._launch_id = self.rp.start_launch(**sl_pt)
@@ -323,15 +349,13 @@ class PyTestServiceClass(object):
         elif leaf['type'] != LeafType.ROOT:
             self._tree_path[leaf['item']] = path + [leaf]
 
+    @check_rp_enabled
     def collect_tests(self, session):
         """
         Collect all tests.
 
         :param session: pytest.Session
         """
-        if self.rp is None:
-            return
-
         # Create a test tree to be able to apply mutations
         test_tree = self._build_test_tree(session)
         self._remove_root_package(test_tree)
@@ -415,10 +439,8 @@ class PyTestServiceClass(object):
         leaf['item_id'] = item_id
         leaf['exec'] = ExecStatus.IN_PROGRESS
 
+    @check_rp_enabled
     def _create_suite_path(self, item):
-        if self.rp is None:
-            return
-
         path = self._tree_path[item]
         for leaf in path[1:-1]:
             if leaf['exec'] != ExecStatus.CREATED:
@@ -656,6 +678,7 @@ class PyTestServiceClass(object):
     def __started(self):
         return self.__unique_id() in self._start_tracker
 
+    @check_rp_enabled
     def start_pytest_item(self, test_item=None):
         """
         Start pytest_item.
@@ -663,7 +686,7 @@ class PyTestServiceClass(object):
         :param test_item: pytest.Item
         :return: item ID
         """
-        if self.rp is None or test_item is None:
+        if test_item is None:
             return
 
         if not self.__started():
@@ -758,6 +781,7 @@ class PyTestServiceClass(object):
         self._lock(leaf['parent'], lambda p: self._proceed_suite_finish(p))
         self._finish_parents(leaf['parent'])
 
+    @check_rp_enabled
     def finish_pytest_item(self, test_item):
         """
         Finish pytest_item.
@@ -765,9 +789,6 @@ class PyTestServiceClass(object):
         :param test_item: pytest.Item
         :return: None
         """
-        if self.rp is None:
-            return
-
         path = self._tree_path[test_item]
         leaf = path[-1]
         self._process_metadata_item_finish(leaf)
@@ -811,18 +832,17 @@ class PyTestServiceClass(object):
         log.debug('ReportPortal - Finish launch: request_body=%s', finish_rq)
         self.rp.finish_launch(**finish_rq)
 
+    @check_rp_enabled
     def finish_launch(self):
         """
         Finish tests launch.
 
         :return: None
         """
-        if self.rp is None:
-            return
-
         # To finish launch session str parameter is needed
         self._finish_launch(self._build_finish_launch_rq())
 
+    @check_rp_enabled
     def post_log(self, test_item, message, log_level='INFO', attachment=None):
         """
         Send a log message to the Report Portal.
@@ -834,9 +854,6 @@ class PyTestServiceClass(object):
         :param attachment: attachment file
         :return: None
         """
-        if self.rp is None:
-            return
-
         if log_level not in self._log_levels:
             log.warning('Incorrect loglevel = %s. Force set to INFO. '
                         'Available levels: %s.', log_level, self._log_levels)
@@ -850,7 +867,7 @@ class PyTestServiceClass(object):
         }
         self.rp.log(**sl_rq)
 
-    def start(self):
+    def start(self) -> None:
         """Start servicing Report Portal requests."""
         self.parent_item_id = self._config.rp_parent_item_id
         self.ignored_attributes = list(