+++ tests/__init__.py
@@ -14,9 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License
 """
 
-from six import add_move, MovedModule
-add_move(MovedModule('mock', 'mock', 'unittest.mock'))
-
 REPORT_PORTAL_PACKAGE = 'pytest_reportportal'
 
 REPORT_PORTAL_SERVICE = REPORT_PORTAL_PACKAGE + '.service.RPClient'