+++ tests/integration/test_pytest_parallel.py
@@ -15,7 +15,7 @@
 
 import pytest
 from delayed_assert import expect, assert_expectations
-from six.moves import mock
+from unittest import mock
 
 from tests import REPORT_PORTAL_SERVICE
 from tests.helpers import utils