+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 ### Added
+- `rp_launch_uuid_print` and `rp_launch_uuid_print_output` configuration parameters, by @HardNorth
+### Removed
+- Python 2.7, 3.6 support, by @HardNorth
+
+## [5.1.9]
+### Added
 - `rp_api_retries` configuration parameter, by @HardNorth
 ### Changed
 - Client version updated on [5.3.5](https://github.com/reportportal/client-Python/releases/tag/5.3.5), by @HardNorth